% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBS_ITC.R
\name{CBS_ITC}
\alias{CBS_ITC}
\title{CBS_ITC.R}
\usage{
CBS_ITC(choice, Amt1, Var1, Amt2, Var2, numpiece)
}
\arguments{
\item{choice}{Vector of 0s and 1s. 1 if the choice was option 1, 0 if the choice was option 2}

\item{Amt1}{Vector of real numbers. Reward amount of choice 1.}

\item{Var1}{Vector of real numbers. Delay until the reward of choice 1. Should be normalized within 0 and 1}

\item{Amt2}{Vector of real numbers. Reward amount of choice 2.}

\item{Var2}{Vector of real numbers. Delay until the reward of choice 2. Should be normalized within 0 and 1}

\item{numpiece}{Either 1 or 2. Number of CBS pieces to use.}
}
\value{
A list containing the following:
\itemize{
    \item \code{type}: either 'CBS1' or 'CBS2' depending on the number of pieces
    \item \code{LL}: log likelihood of the model
    \item \code{numparam}: number of total parameters in the model
    \item \code{scale}: scaling factor of the logit model
    \item \code{xpos}: x coordinates of the fitted CBS function
    \item \code{ypos}: y coordinates of the fitted CBS function
    \item \code{AUC}: area under the curve of the fitted CBS function
    \item \code{Origmodel} : NlcOptim object of the fitted CBS function. Not intended to be useful, but contains diagnostic information such as convergence condition, iterations, etc.
}
}
\description{
Fit either a 1-piece or 2-piece CBS latent utility function to binary intertemporal choice data.
}
\details{
The data should be consisted of n trials (ideally n > 100) in which a participant made choices between two reward options.
At least one of the two options should be delayed such that option 1 is receiving \code{Amt1} in delay1 and option 2 is receiving
\code{Amt2} in delay2. If one of the options are immediate, than delay may be 0. If the participant chose option 1, \code{choice}
should be 1 and 0 otherwise. Note that before running this function, the delay should be normalized to [0 1]. For example if
delay1 ranges from 0 to max_delay1, and delay2 ranges from 20 to max_delay2, one could normalize all delays by \code{normD = max{max_delay1,max_delay2}}.
Hence, \code{Var1} should be delay1/normD and \code{Var2} should be delay2/normD. This normalization is not done internally because
the choice of the normalization constant affects how the CBS function can fit the data.
}
